Creacion de controladores JobPost, Admin

+++ backend/src/main/java/com/webAppG9/backend/repository/UserRepository.java
@@ -3,16 +3,11 @@ package com.webAppG9.backend.repository;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.webAppG9.backend.Model.User;
 
-//dmetodo interfaz qeu  hereda todos los metodos de JraRepository
-//trabaja con la entidad User y el tipo de ID (Long)
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    // Optional para que en caso de no halklar nada no de null,
+    // Buscar un usuario por email (opcional)
     Optional<User> findByEmail(String email);
 
-    // buscar pasanndo el tipo de dato y el valor de la entidad
-    // User findByUsername(String username);
 }
